# server/docs/swagger.yaml
basePath: /
definitions:
...
  application.App:
    properties:
      createdAt:
        description: 创建时间
        type: string
      id:
        description: 主键ID
        type: integer
      name:
        description: 应用名称
        type: string
      type:
        description: 应用类型 0 未定义 1 数据库 2 缓存 3 应用 4 存储 5 负载均衡 6 备份 7 反向代理
        type: integer
      updatedAt:
        description: 更新时间
        type: string
      version:
        description: 版本
        type: string
    type: object
  application.ApplicationServer:
    properties:
      appIds:
        description: 安装应用列表
        type: string
      apps:
        description: 安装应用列表
        items:
          $ref: '#/definitions/application.App'
        type: array
      architecture:
        description: 架构
        type: integer
      createdAt:
        description: 创建时间
        type: string
      hostname:
        description: 主机名
        type: string
      id:
        description: 主键ID
        type: integer
      manageIp:
        description: 管理ip
        type: string
      os:
        description: 系统
        type: integer
      osVersion:
        description: 系统版本
        type: string
      sshPort:
        description: ssh端口
        type: integer
      systemId:
        description: 所属系统id
        type: integer
      updatedAt:
        description: 更新时间
        type: string
    type: object
...
  request.GetById:
    properties:
      id:
        description: 主键ID
        type: number
      page:
        description: 页码
        type: integer
      pageSize:
        description: 每页大小
        type: integer
      sortBy:
        type: string
      sortInverted:
        type: boolean
    type: object
...
  request.IdsReq:
    properties:
      ids:
        items:
          type: integer
        type: array
    type: object
...
  request.PageInfo:
    properties:
      page:
        description: 页码
        type: integer
      pageSize:
        description: 每页大小
        type: integer
    type: object
...
  request.TemplateScriptRequest:
    properties:
      detail:
        type: boolean
      id:
        description: 主键ID
        type: number
      serverId:
        type: number
    type: object
...
  taskMdl.Task:
    properties:
      beginTime:
        description: 开始时间
        type: string
      createdAt:
        description: 创建时间
        type: string
      endTime:
        description: 结束时间
        type: string
      id:
        description: 主键ID
        type: integer
      status:
        description: 状态
        type: string
      systemUserId:
        description: 执行人
        type: integer
      templateId:
        description: task id
        type: integer
      updatedAt:
        description: 更新时间
        type: string
    type: object
  taskMdl.TaskTemplate:
    properties:
      command:
        description: 命令
        type: string
      createdAt:
        description: 创建时间
        type: string
      description:
        description: task描述
        type: string
      id:
        description: 主键ID
        type: integer
      lastTask:
        $ref: '#/definitions/taskMdl.Task'
      lastTaskId:
        description: 最后一次task id
        type: integer
      mode:
        description: 执行方式 1 命令 2 脚本
        type: integer
      name:
        description: task名称
        type: string
      scriptPath:
        description: 脚本位置
        type: string
      sysUser:
        description: 执行用户
        type: string
      targetIds:
        items:
          type: integer
        type: array
      targetServerIds:
        description: 关联服务器id
        type: string
      targetServers:
        items:
          $ref: '#/definitions/application.ApplicationServer'
        type: array
      updatedAt:
        description: 更新时间
        type: string
    type: object
info:
  contact: {}
  description: This is a sample Server pets
  title: Swagger Example API
  version: 0.0.1
paths:
...
  /task/template/addTemplate:
    post:
      consumes:
      - application/json
      parameters:
      - description: Template
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/taskMdl.TaskTemplate'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"添加成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 新增Template
      tags:
      - Template
  /task/template/checkScript:
    post:
      consumes:
      - application/json
      parameters:
      - description: 主机名, 架构, 管理ip, 系统, 系统版本
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.TemplateScriptRequest'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"更新成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 检查script
      tags:
      - Template
  /task/template/deleteTemplate:
    post:
      consumes:
      - application/json
      parameters:
      - description: TaskTemplateId
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.GetById'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"删除成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 删除Template
      tags:
      - Template
  /task/template/deleteTemplateByIds:
    post:
      consumes:
      - application/json
      parameters:
      - description: ID
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.IdsReq'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"删除成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 批量删除Template
      tags:
      - Template
  /task/template/downloadScript:
    get:
      consumes:
      - application/json
      parameters:
      - description: 主机名, 架构, 管理ip, 系统, 系统版本
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.TemplateScriptRequest'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"下载成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 下载script
      tags:
      - Template
  /task/template/getTemplateById:
    post:
      consumes:
      - application/json
      parameters:
      - description: TemplateId
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.GetById'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 根据id获取Template
      tags:
      - Template
  /task/template/getTemplateList:
    post:
      consumes:
      - application/json
      parameters:
      - description: 页码, 每页大小
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.PageInfo'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 分页获取基础Template列表
      tags:
      - Template
  /task/template/updateTemplate:
    post:
      consumes:
      - application/json
      parameters:
      - description: 主机名, 架构, 管理ip, 系统, 系统版本
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/taskMdl.TaskTemplate'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"更新成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 更新Template
      tags:
      - Template
  /task/template/uploadScript:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: 导入Script文件
        in: formData
        name: file
        required: true
        type: file
      - description: int valid
        in: query
        name: int
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"导入成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 导入脚本
      tags:
      - Template
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"